Refactor "controllers" to "services"
